Improved google drive support, added downloadable text type
